+++ lib/sidekiq/redis_connection.rb
@@ -9,7 +9,7 @@ module Sidekiq
       driver = options[:driver] || 'ruby'
       # need a connection for Fetcher and Retry
       size = options[:size] || (Sidekiq.server? ? (Sidekiq.options[:concurrency] + 2) : 5)
-      namespace = options[:namespace] || Sidekiq.options[:namespace]
+      namespace = options[:namespace]
 
       ConnectionPool.new(:timeout => 1, :size => size) do
         build_client(url, namespace, driver)